requirements.txt
absl-py==2.3.1
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alabaster
altair
anaconda-anon-usage
anaconda-auth
anaconda-catalogs
anaconda-cli-base
anaconda-client
anaconda-navigator
anaconda-project
annotated-types==0.7.0
anyio
appdirs==1.4.4
archspec
argon2-cffi
argon2-cffi-bindings
arrow
asgiref==3.9.1
astroid
astropy
astropy-iers-data
asttokens
astunparse==1.6.3
async-lru
asyncssh
atomicwrites==1.4.0
attrs==25.4.0
Automat
autopep8
babel
bcrypt
beautifulsoup4
binaryornot
black==25.1.0
bleach
blinker
bokeh
boltons
boto3==1.40.39
botocore==1.40.59
Bottleneck
Brotli
cachetools
cattrs==25.3.0
certifi
cffi
cfgv==3.4.0
chardet
charset-normalizer
click==8.2.1
cloudpickle
colorama==0.4.6
colorcet
comm
conda
conda-anaconda-telemetry
conda-anaconda-tos
conda-build
conda-content-trust
conda-libmamba-solver
conda-pack
conda-package-handling
conda-repo-cli
conda-token
conda_index
conda_package_streaming
constantly
contourpy
cookiecutter
cryptography
cssselect
cycler
cytoolz
dask
dask-expr
datasets==4.1.1
datashader
debugpy
decorator
defusedxml
Deprecated
diff-match-patch
dill
distlib==0.4.0
distributed
distro
docstring-to-markdown
docutils
drf-spectacular==0.28.0
et_xmlfile
evalidate
evaluate==0.4.6
executing
factory-boy==3.3.0
Faker==37.12.0
fastjsonschema
filelock==3.19.1
flake8
Flask
flatbuffers==25.2.10
fonttools
frozendict
frozenlist
fsspec
gast==0.6.0
gitdb
GitPython
gmpy2
gnome-extensions-cli==0.10.6
google-pasta==0.2.0
greenlet
grpcio==1.75.0
h11
h5py
HeapDict
hf-xet==1.1.10
holoviews
httpcore
httpx
huggingface-hub==0.35.0
hvplot
hyperlink
identify==2.6.13
idna
imageio
imagesize
imbalanced-learn
importlib_metadata
incremental
inflection
iniconfig
intake
intervaltree
ipykernel
ipython
ipywidgets
isort==6.0.1
itemadapter
itemloaders
itsdangerous
jaraco.classes
jaraco.context
jaraco.functools
jedi
jeepney
jellyfish
Jinja2
jinja2-time
jiter==0.11.0
jmespath
joblib
json5
jsonpatch
jsonpointer==2.1
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kagglehub==0.3.13
keras==3.11.3
keyring
kiwisolver
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.31
lazy_loader
lckr_jupyterlab_variableinspector
libarchive-c
libclang==18.1.1
libmambapy
lief
linkify-it-py
llvmlite
lmdb
locket
lsprotocol==2025.0.0
lxml
lz4
Markdown
markdown-it-py
MarkupSafe
matplotlib==3.10.0
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
menuinst
mistune
mkl-service==2.4.0
mkl_fft
mkl_random
ml_dtypes==0.5.3
more-itertools
mpmath
msgpack
multidict
multipledispatch
multiprocess==0.70.16
mypy==1.17.1
mypy_extensions==1.1.0
namex==0.1.0
narwhals
navigator-updater
nbclient
nbconvert
nbformat
nest_asyncio
networkx
nltk
nodeenv==1.9.1
notebook
notebook_shim
numba
numexpr
numpy
numpydoc
openai==1.108.1
openpyxl
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
overrides
packaging
pandas
pandocfilters
panel
param
parsel
parso
partd
pathspec==0.12.1
patsy
pexpect
pickleshare
pillow
pkce
pkginfo
platformdirs==4.3.8
plotly
pluggy
ply
poyo
pre_commit==4.3.0
prometheus_client
prompt_toolkit
propcache
Protego
protobuf
psutil
psycopg2-binary==2.9.10
ptyprocess
pure-eval
py-cpuinfo
pyarrow==21.0.0
pyasn1
pyasn1-modules==0.2.8
pycodestyle
pycosat
pycparser
pyct
pycurl
pydantic==2.11.7
pydantic-settings
pydantic_core==2.33.2
PyDispatcher
pydocstyle
pyerfa
pyflakes
PyGithub
Pygments
PyJWT
pylint
pylint-venv
pyls-spyder==0.4.0
pylsp-mypy==0.7.0
PyNaCl
pyodbc
pyOpenSSL
pyparsing
PyQt5==5.15.10
PyQt5_sip
PyQtWebEngine==5.15.6
PySocks
pytest==8.0.0
pytest-django==4.8.0
python-dateutil
python-dotenv
python-json-logger
python-lsp-black
python-lsp-jsonrpc
python-lsp-ruff==2.2.2
python-lsp-server==1.13.0
python-slugify
pytoolconfig
pytz==2025.2
pyuca
pyviz_comms
PyWavelets
pyxdg
PyYAML
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
queuelib
readchar
referencing
regex
requests
requests-file
requests-toolbelt
rfc3339_validator
rfc3986_validator
rich
roman-numerals-py
rope
rpds-py
Rtree
ruamel.yaml
ruamel.yaml.clib
ruamel_yaml_conda
ruff==0.12.10
s3fs
s3transfer==0.14.0
safetensors==0.6.2
scikit-image
scikit-learn
scipy
Scrapy
seaborn
SecretStorage
semver
Send2Trash
service-identity
setuptools==72.1.0
shellingham
sip
six
sklearn-compat
smmap
sniffio
snowballstemmer
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
sqlparse==0.5.3
stack-data
statsmodels
streamlit
superqt
sympy
tables
tabulate
tblib
tenacity
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado
text-unidecode
textdistance
tf_keras==2.20.1
threadpoolctl
three-merge
tifffile
tiktoken==0.11.0
tinycss2
tldextract
tokenizers==0.22.1
toml
tomli
tomlkit
toolz
tornado
tqdm==4.67.1
traitlets
transformers==4.56.2
truststore
Twisted
typer
types-PyYAML==6.0.12.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata
uc-micro-py
ujson
Unidecode
uritemplate==4.2.0
urllib3
virtualenv==20.34.0
w3lib
watchdog
wcwidth
webencodings
websocket-client
Werkzeug
whatthepatch
wheel==0.45.1
widgetsnbextension
wrapt
wurlitzer
xarray
xxhash==3.5.0
xyzservices
yapf
yarl
zict
zipp
zope.interface
zstandard
langchain_google_genai-3.2.0
